Move TS analyis code to its own place
