+++ src/staticAnalysis/tsDiagUtil.ts
